+++ modules/tinymce/CHANGELOG.md
@@ -56,7 +56,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved the `hr` plugin's functionality to TinyMCE core #TINY-8313
 - Moved the `print` plugin's functionality to TinyMCE core #TINY-8314
 - Renamed the `getShortEndedElements` Schema API to `getVoidElements` #TINY-8344
-- Renamed `font_formats` to `font_family_formats`, `fontselect` to `fontfamily`, and `fontformats` to `fontfamily` #TINY-8328
+- Renamed `font_formats` to `font_family_formats`, `fontselect` toolbar button to `fontfamily`, and `fontformats` menu item to `fontfamily` #TINY-8328
+- Renamed `formatselect` toolbar button and `blockformats` menu item to `blocks` #TINY-8328
+
 
 ### Fixed
 - The object returned from the `editor.fire()` API was incorrect if the editor had been removed #TINY-8018